[refactor] recruitPage, reviewPage Oultet을 통해 공통 컴포넌트 분리
