+++ src/Router/routes.tsx
@@ -10,7 +10,8 @@ import PrivateRoutes from "../Components/Layout/PrivateRoutes";
 import PublicOnlyRoutes from "../Components/Layout/PublicOnlyRoutes";
 import Layout from "../Components/Layout";
 import { ProfileRequiredRoutes } from "../Components/Layout/ProfileRequiredRoutes";
-import MypageTest from '../Components/Mypage/test/MypageTest';
+import MypageTest from "../Components/Mypage/test/MypageTest";
+import HomeLayout from "../Components/Layout/HomeLayout";
 
 const Error = React.lazy(() => import("../Page/Error"));
 const SignIn = React.lazy(() => import("../Page/SignIn"));
@@ -56,14 +57,17 @@ const Router = (): JSX.Element => {
               <Route path="travelHistory" element={<MyTravelHistory />} />
               <Route path="accountSettings" element={<MyAccountSettings />} />
             </Route>
-              <Route path="test" element={<MypageTest />} />
+            <Route path="test" element={<MypageTest />} />
           </Route>
 
           {/* 로그인 필요없는 일반 페이지 */}
+          <Route element={<HomeLayout />}>
+            <Route path="/recruitment" element={<Recruitment />} />
+            <Route path="/review" element={<Review />} />
+          </Route>
+
           <Route path="/" element={<Navigate to="/recruitment" replace />} />
-          <Route path="/recruitment" element={<Recruitment />} />
           <Route path="/recruitment/:id" element={<RecruitPost />} />
-          <Route path="/review" element={<Review />} />
           <Route path="/review/:id" element={<ReviewPost />} />
           <Route path="*" element={<Error />} />
         </Route>
